cloud_function/requirements.txt
functions-framework==3.5.0
flask==3.0.0

# Google Cloud
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.73.0

# Agent dependencies
langgraph==0.2.28
langchain-google-vertexai==2.0.3
langchain-core==0.3.15

# GitHub
PyGithub==2.4.0

# Code analysis
ast-grep-py==0.25.4
libcst==1.4.0

# Utilities
pydantic==2.9.2
requests==2.32.3
